Move Target GUIs into Targeting namespace
