+++ src/RealAntennasProject/RealAntennas.csproj
@@ -37,7 +37,7 @@
     <LangVersion>8.0</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Antenna\AntennaGUI.cs" />
+    <Compile Include="Targeting\AntennaTargetGUI.cs" />
     <Compile Include="Antenna\BandInfo.cs" />
     <Compile Include="Antenna\Encoder.cs" />
     <Compile Include="Kerbalism\Kerbalism.cs" />
@@ -68,7 +68,7 @@
     <Compile Include="Network\ScenarioPatcher.cs" />
     <Compile Include="RealAntennasUI.cs" />
     <Compile Include="MapUI\SignalToolTipController.cs" />
-    <Compile Include="RemoteAntennaControlUI.cs" />
+    <Compile Include="Targeting\RemoteAntennaControlUI.cs" />
     <Compile Include="Targeting\AntennaTarget.cs" />
     <Compile Include="Targeting\AntennaTargetManager.cs" />
     <Compile Include="TechLevelInfo.cs" />
